Add embed handler, improve api handler and fix eval command.

+++ cmds/hdposter.js
@@ -13,12 +13,11 @@ c.process = async (bot, msg, cmdArgs, guild, user, config, u) => {
     let message = await bot.createMessage(msg.channel.id, `ℹ Getting HD poster for the title with IMDb ID '**${cmdArgs[0]}**'...`);
     
     let poster = await u.api.getHDPoster(cmdArgs[0]);
-    if (poster && poster.Error) return message.edit(`❌ ${poster.Error}`);
-    if (!poster) return message.edit('❌ Unknown Error.');
+    if (!poster || poster && poster.Error) return message.edit(`❌ ${poster.Error || 'Uknown Error.'}`);
 
     message.delete();
     bot.createMessage(msg.channel.id, '', {
-        'file': poster.data,
+        'file': poster,
         'name': cmdArgs[0] + '.png'
     });
 }
